requirements.txt
AWSIoTPythonSDK
inky
font_fredoka_one
Pillow
boto3